Add backup module database for bootstrapping purposes
Populating a new database, at this point in the game, requires
access to the old one.

SQLite doesn't make migrations easy, and we'll want to build a
leaner versioned-release-only database for public use.

For now, this doesn't do anything fancy to find the bootstrap
database, since I envision this as a core-developer tool.

Nothing untoward should happen if ~bridge.modules doesn't exist,
it doesn't for me.