Use "rw" to open modules
The *default* behavior is to open in "rwc" mode, creating the
database if it doesn't exist.

We pcall the open and assign if it doesn't fail.

This commit also moves the custom loaders behind the package.preload
loader.  We'll eventually replace all of this with a custom
implementation (`use`), but in the meantime, this better expresses
the intention.