Coerce boolean values to 0 and 1
This is the expected convention for storing booleans in a SQLite
database.

No reverse coercion is performed, so the application will have to
keep track of what values are expected to be boolean and turn them
back into bools manually.

A column may be constrained at the database level to only accept
booleans using a CHECK constraint.