Make "k" the default for resultset
I'm not convinced this is the better thing to do, but nearly so.

I end up using "i" more than "k" but this is also a bad use of
SQL: I nearly always use an 'unwrap key' helper function, when
what I really want is a special SQL verb that returns either one
result or `nil`, for when I know that I'm only going to get one
result from one column.

Usually a key, but not often enough that I want to commit to that.