un-JIT sql.orb
This appears to fix the problem, so my intuition that we're dealing
with an intersection of the SQLite code and the JIT has some more
backing.

It would of course be nice to have the SQLite code JITted, but
surely it spends most of its time on syscalls anyway.

Hope this fixes the problem, for now.