Use sqlite3_bind_parameter_name
Instead of iterating all the keys of the provided table, we
retrieve the names of the parameters, and bind any which match the
names of keys in the table.

This is O(m) where m is the number of parameters in the prepared
statement, instead of O(n) where n is the number of keys in the
table, and m < n for normal workloads.