Return index in stmt:cols()
Lua iterators shut down whenever the first of the returned values
is a nil.

Since sql.orb represents NULLs as `nil` (usually good, sometimes
problematic, like here), and column values can of course be null,
this is a latent bug.

Returning the index is useful sometimes anyway, and I was using
this in two places before I caught it, so the fix is relatively
painless.