Add shim for running custom verbs
This is a bit of an (untested) hack, but it should serve to allow
projects to run from the command line, including passing through
remaining arguments to a custom argparse.

We'll continue to parse the core tools from pylon, to allow all the
affordances like help menus and autocompletions for bash/zsh/fish
to work correctly.