SQLite busy retry... for conn:prepare
This works in async and blocking mode for conn:prepare, because I
can get the database to block before preparing a statement.

But I haven't been able to figure out how to get it to block on
a step. so the step code probably works for blocking, but I didn't
even bother writing the async block code, and won't unless I can
figure out how to simulate contention.

I'll take it up with the SQLite forum, I guess.