Reject integer session names
We don't want any sessions to have integer names, because we offer
the ability to refer to a session by number in as many places as
possible, and having a session called "4" with the number 6 means
one would only ever be able to refer to it by that number.

I had to do something kind of janky to get different validations on
`br session rename 'old' 'new'`, but it's not the end of the world.